pyproject.toml
[project]
name = "vea"
version = "0.1.0"
description = "Generates a daily briefing from your Google Calendar, Gmail, Markdown journal files, Slack messages, and Todoist task list."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name="Floris Vlasveld", email="floris@floris.io" }
]
dependencies = [
  "openai>=1.14",
  "google-auth>=2.29",
  "google-api-python-client>=2.125",
  "google-auth-oauthlib>=1.0.0",
  "todoist-api-python>=3",
  "python-dotenv>=0.21",
  "typer>=0.6.1",
  "slack-sdk>=3.19",
  "rich>=13.0",
  "pytz>=2022.1",
  "anthropic>=0.1.3",
  "google-generativeai>=0.8.5",
  "weasyprint>=65.1",
  "markdown>=3.8"
  ,"sentence-transformers>=2.2.2"
  ,"faiss-cpu>=1.7.4"
]

[project.scripts]
vea = "vea.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["vea*"]
exclude = ["credentials*"]

[tool.ruff]
line-length = 120
target-version = "py310"
